+++ src/routes/graphql/profile/profileQuery.ts
@@ -1,7 +1,7 @@
 import { GraphQLList, GraphQLObjectType } from 'graphql';
 import { UUIDType } from '../types/uuid.js';
 import { Context } from '../types/context.js';
-import { ProfileType } from './typeProfile.js';
+import { ProfileType } from './profileType.js';
 import { Profile } from '@prisma/client';
 
 export const getProfileQuery = {
