+++ src/routes/graphql/users/userType.ts
@@ -1,9 +1,9 @@
 import { GraphQLFloat, GraphQLList, GraphQLObjectType, GraphQLString } from 'graphql';
 import { UUIDType } from '../types/uuid.js';
-import { ProfileType } from '../profile/typeProfile.js';
+import { ProfileType } from '../profile/profileType.js';
 import { User } from '@prisma/client';
 import { Context, UserSub } from '../types/context.js';
-import { PostType } from '../post/typePost.js';
+import { PostType } from '../post/postType.js';
 
 export const UserType = new GraphQLObjectType({
   name: 'User',
@@ -75,4 +75,3 @@ export const UserType = new GraphQLObjectType({
     },
   }),
 });
-0 comments on comm
